update baiduyun
1. updated to support upload specail file of types.
2. updated bakmypy.py to upload source files.

+++ tools/bakmypy.py
@@ -11,7 +11,11 @@ import shutil
 import tarfile
 import time
 
+from mypy.pybase import print_help, print_exit
 from baiduyun import BaiduYun
+
+VERSION = '1.2.0'
+
 
 MYPY = os.path.join(os.getenv('HOME'), 'mypy')
 
@@ -36,10 +40,40 @@ def create_tar(path):
                     fd.add(f)
     return tar
 
+
 if __name__ == '__main__':
-    clean_cache(MYPY)
-    tar = create_tar(MYPY)
-    args = {'-l': tar, '-y': 'Mypy', '-c': 'upload'}
-    by = BaiduYun('BackupMypy', True)
-    by.main(args)
-    os.remove(tar)
+    from mypy.pybase import get_user_input
+
+    HELP_MENU = (
+        '============================================',
+        '    bakmypy help',
+        '============================================',
+        'options:',
+        ' -a: upload all of files to baiduyun.',
+        ' -f: upload source files to baiduyun.',
+        ' -z: tar files to mypy_(date-time).bz2 and upload to baiduyun.',
+
+    )
+
+    args = get_user_input('hafz')
+    if args:
+        dt = time.strftime('%Y%m%d-%H%M', time.localtime())
+        clean_cache(MYPY)  # clear all of cache files: *.pyc and __pyc__ dirs
+        by = BaiduYun('BackupMypy')
+        if '-h' in args:
+            print_help(HELP_MENU)
+        elif '-a' in args:
+            args = {'-l': MYPY, '-y': 'Mypy/mypy_%s' % dt, '-c': 'upload'}
+            by.main(args)
+        elif '-f' in args:
+            args = {'-l': MYPY, '-y': 'Mypy/mypy_%s' % dt,
+                    '-f': ('.py', '.txt'), '-c': 'upload'}
+            by.main(args)
+        elif '-z' in args:
+            tar = create_tar(MYPY)
+            args = {'-l': tar, '-y': 'Mypy', '-c': 'upload'}
+            by.main(args)
+            os.remove(tar)
+    else:
+        print_exit('no input, -h for help')
+
